Tindeer
Self made website.
I tried to build the website purely with knowledge I have after taking section of a course in bootcamp.
Tried my best to search for information I needed to make the website if I forgot any concepts. 

Its not much but I would say this is a good display of what I'm currently capable of making for now. 
Obviously it's not perfect but I think this is good starting point and hopefully, I'll get much better and come back to this project to expand on it.